Adds venueGetEvents and separates the test.
